Init bindings templates
Use the default templates available in the Github repository
of the openapi-python-client package:
https://github.com/openapi-generators/openapi-python-client/tree/v.0.10.7/openapi_python_client/templates

Use explicitly v.0.10.7 tag from which the templates will be modified
later the same way as they are modified in the osidb-bindings project.
Since the templates and openapi-python-client API for the template
generation is a beta feature and it is unstable, we need to lock the
version we are using.