Replace httpx with requests
Overall the requests library is just much more mature and secure and
recent dependabot scans revelaed some very serious issues with the
httpx library, so we are shifting from to the requests.

Note: httpx will stay only in devel-requirements as it is a dependency
of the openapi-python-client package which is used to build the
bindings. They only use the httpx in some test cases and for demo
purposes so nothing that should really bother us.