Add empty value to all enums
This is a workaround for the current status of Component Registry since
it provides empty (NONE) values at some enums fields which does not
allow empty values and thus bindings then fails to receive the response.

Once these enums are resolved on the Component Registry side, we can
safely remove this compatibility shim.